Add parser functionality for add_tag command
