--- src/main/java/seedu/address/model/ReadOnlyTagList.java
@@ -1,16 +1,0 @@
-package seedu.address.model;
-
-import javafx.collections.ObservableList;
-import seedu.address.model.tag.Tag;
-
-/**
- * Unmodifiable view of a tags collection
- */
-public interface ReadOnlyTagList {
-
-    /**
-     * Returns an unmodifiable view of tags collection.
-     * This list will not contain any duplicate tags.
-     */
-    ObservableList<Tag> getTagList();
-}
