--- src/main/java/seedu/address/model/TagList.java
@@ -1,106 +1,0 @@
-package seedu.address.model;
-
-import static java.util.Objects.requireNonNull;
-
-import java.util.List;
-
-import javafx.collections.ObservableList;
-import seedu.address.commons.util.ToStringBuilder;
-import seedu.address.model.tag.Tag;
-import seedu.address.model.tag.UniqueTagList;
-
-public class TagList implements ReadOnlyTagList {
-
-    private final UniqueTagList tags;
-
-    {
-        tags = new UniqueTagList();
-    }
-
-    public TagList() {}
-
-    public TagList(ReadOnlyTagList toBeCopied) {
-        this();
-        resetData(toBeCopied);
-    }
-
-    /**
-     * Replaces the contents of the tags list with {@code tags}.
-     * {@code tags} must not contain duplicate tags.
-     */
-    public void setTags(List<Tag> tags) {
-        this.tags.setTags(tags);
-    }
-
-    /**
-     * Resets the existing data of this {@code TagList} with {@code newData}.
-     */
-    public void resetData(ReadOnlyTagList newData) {
-        requireNonNull(newData);
-
-        setTags(newData.getTagList());
-    }
-
-    //// tag-level operations
-
-    /**
-     * Returns true if a tag with the same identity as {@code tag} exists in the address book.
-     */
-    public boolean hasTag(Tag tag) {
-        requireNonNull(tag);
-        return tags.contains(tag);
-    }
-
-    /**
-     * Adds a tag to the address book.
-     * The tag must not already exist in the address book.
-     */
-    public void addTag(Tag p) {
-        tags.add(p);
-    }
-
-    /**
-     * Replaces the given tag {@code target} in the list with {@code editedTag}.
-     * {@code target} must exist in the address book.
-     * The tag identity of {@code editedTag} must not be the same as another existing tag in the address book.
-     */
-    public void setTag(Tag target, Tag editedTag) {
-        requireNonNull(editedTag);
-
-        tags.setTag(target, editedTag);
-    }
-
-    //// util methods
-
-    @Override
-    public String toString() {
-        return new ToStringBuilder(this)
-                .add("tags", tags)
-                .toString();
-    }
-
-    @Override
-    public ObservableList<Tag> getTagList() {
-        return tags.asUnmodifiableObservableList();
-    }
-
-    @Override
-    public boolean equals(Object other) {
-        if (other == this) {
-            return true;
-        }
-
-        // instanceof handles nulls
-        if (!(other instanceof TagList)) {
-            return false;
-        }
-
-        TagList otherTagList = (TagList) other;
-        return tags.equals(otherTagList.tags);
-    }
-
-    @Override
-    public int hashCode() {
-        return tags.hashCode();
-    }
-}
